requirements.txt
confluent_kafka
urllib3
requests
avro
matplotlib
paho-mqtt